# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi Swagger untuk backend be_latihan menggunakan Golang Fiber, GORM, PostgreSQL, dan JWT Bearer Token.",
        "title": "API be_latihan Praktikum 13",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Praktikum Pemrograman III",
            "email": "praktikum@example.com"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:3000",
    "basePath": "/",
    "paths": {
        "/api/mahasiswa/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil seluruh data mahasiswa. Endpoint ini hanya dapat diakses oleh admin.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Ambil semua data mahasiswa",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data mahasiswa",
                        "schema": {
                            "$ref": "#/definitions/model.MahasiswaListSuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Token tidak dikirim atau tidak valid",
                        "schema": {
                            "$ref": "#/definitions/model.UnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "User tidak memiliki akses admin",
                        "schema": {
                            "$ref": "#/definitions/model.ForbiddenResponse"
                        }
                    },
                    "500": {
                        "description": "Gagal mengambil data mahasiswa",
                        "schema": {
                            "$ref": "#/definitions/model.InternalServerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menambahkan data mahasiswa baru. Endpoint ini hanya dapat diakses oleh admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Tambah data mahasiswa",
                "parameters": [
                    {
                        "description": "Data mahasiswa baru",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Mahasiswa"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Berhasil menambahkan data mahasiswa",
                        "schema": {
                            "$ref": "#/definitions/model.MahasiswaCreatedSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Payload tidak valid",
                        "schema": {
                            "$ref": "#/definitions/model.BadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Token tidak dikirim atau tidak valid",
                        "schema": {
                            "$ref": "#/definitions/model.UnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "User tidak memiliki akses admin",
                        "schema": {
                            "$ref": "#/definitions/model.ForbiddenResponse"
                        }
                    },
                    "500": {
                        "description": "Gagal menambahkan data mahasiswa",
                        "schema": {
                            "$ref": "#/definitions/model.InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/mahasiswa/{npm}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil satu data mahasiswa berdasarkan NPM. Endpoint ini hanya dapat diakses oleh admin.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Ambil detail mahasiswa berdasarkan NPM",
                "parameters": [
                    {
                        "type": "string",
                        "description": "NPM mahasiswa",
                        "name": "npm",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data mahasiswa",
                        "schema": {
                            "$ref": "#/definitions/model.MahasiswaDetailSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "NPM tidak valid",
                        "schema": {
                            "$ref": "#/definitions/model.BadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Token tidak dikirim atau tidak valid",
                        "schema": {
                            "$ref": "#/definitions/model.UnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "User tidak memiliki akses admin",
                        "schema": {
                            "$ref": "#/definitions/model.ForbiddenResponse"
                        }
                    },
                    "404": {
                        "description": "Data mahasiswa tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/model.NotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Gagal mengambil data mahasiswa",
                        "schema": {
                            "$ref": "#/definitions/model.InternalServerErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengubah data mahasiswa berdasarkan NPM. Endpoint ini hanya dapat diakses oleh admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Ubah data mahasiswa",
                "parameters": [
                    {
                        "type": "string",
                        "description": "NPM mahasiswa",
                        "name": "npm",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data mahasiswa yang diperbarui",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Mahasiswa"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengubah data mahasiswa",
                        "schema": {
                            "$ref": "#/definitions/model.MahasiswaUpdatedSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "NPM atau payload tidak valid",
                        "schema": {
                            "$ref": "#/definitions/model.BadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Token tidak dikirim atau tidak valid",
                        "schema": {
                            "$ref": "#/definitions/model.UnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "User tidak memiliki akses admin",
                        "schema": {
                            "$ref": "#/definitions/model.ForbiddenResponse"
                        }
                    },
                    "404": {
                        "description": "Data mahasiswa tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/model.NotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Gagal mengubah data mahasiswa",
                        "schema": {
                            "$ref": "#/definitions/model.InternalServerErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus data mahasiswa berdasarkan NPM. Endpoint ini hanya dapat diakses oleh admin.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mahasiswa"
                ],
                "summary": "Hapus data mahasiswa",
                "parameters": [
                    {
                        "type": "string",
                        "description": "NPM mahasiswa",
                        "name": "npm",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil menghapus data mahasiswa",
                        "schema": {
                            "$ref": "#/definitions/model.MahasiswaDeletedSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "NPM tidak valid",
                        "schema": {
                            "$ref": "#/definitions/model.BadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Token tidak dikirim atau tidak valid",
                        "schema": {
                            "$ref": "#/definitions/model.UnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "User tidak memiliki akses admin",
                        "schema": {
                            "$ref": "#/definitions/model.ForbiddenResponse"
                        }
                    },
                    "500": {
                        "description": "Gagal menghapus data mahasiswa",
                        "schema": {
                            "$ref": "#/definitions/model.InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.BadRequestResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "detail error validasi"
                },
                "message": {
                    "type": "string",
                    "example": "payload tidak valid"
                }
            }
        },
...
        "model.ForbiddenResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "user tidak memiliki akses untuk fitur ini"
                }
            }
        },
        "model.InternalServerErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "detail error server"
                },
                "message": {
                    "type": "string",
                    "example": "terjadi kesalahan pada server"
                }
            }
        },
...
        "model.Mahasiswa": {
            "type": "object",
            "properties": {
                "alamat": {
                    "type": "string",
                    "example": "Bandung"
                },
                "email": {
                    "type": "string",
                    "example": "ahmad.fauzi@example.com"
                },
                "hobi": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Membaca",
                        "Menulis"
                    ]
                },
                "nama": {
                    "type": "string",
                    "example": "Ahmad Fauzi"
                },
                "no_hp": {
                    "type": "string",
                    "example": "081234567890"
                },
                "npm": {
                    "type": "string",
                    "example": "714240047"
                },
                "prodi": {
                    "type": "string",
                    "example": "Teknik Informatika"
                }
            }
        },
        "model.MahasiswaCreatedSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.Mahasiswa"
                },
                "message": {
                    "type": "string",
                    "example": "berhasil menambahkan data mahasiswa"
                }
            }
        },
        "model.MahasiswaDeletedSuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "berhasil menghapus data mahasiswa"
                }
            }
        },
        "model.MahasiswaDetailSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.Mahasiswa"
                },
                "message": {
                    "type": "string",
                    "example": "berhasil mengambil data mahasiswa"
                }
            }
        },
        "model.MahasiswaListSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Mahasiswa"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "berhasil mengambil data mahasiswa"
                }
            }
        },
        "model.MahasiswaUpdatedSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.Mahasiswa"
                },
                "message": {
                    "type": "string",
                    "example": "berhasil mengubah data mahasiswa"
                }
            }
        },
        "model.NotFoundResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "data mahasiswa tidak ditemukan"
                }
            }
        },
...
        "model.UnauthorizedResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "authorization header wajib diisi"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}